Refactoring mostly complete, needs testing
